expansion-signal-detector/requirements.txt
anthropic>=0.20.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.30.0
slack-sdk>=3.20.0
